+++ manage-engine-service-desk-plus/release_notes.md
@@ -1,2 +1,13 @@
 #### What's Improved
-- Added OAuth authentication support.
+
+- Added following actions:
+  - Get Request Task
+  - Get List Request Task
+  - Delete Request Task
+  - Add Request Task
+  - Edit Request Task
+  - Get Request Note
+  - Get List Request Note
+  - Get Task Comments
+  - List Attachments
+  - Download Attachment